requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytube3==9.6.4
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.2
